Fix paths in manifest-task: fails when space in absolute path
When running the build task one of the options that includes building the manifest (e.g. `yeoman build:basics`) the build fails if you got a space in the absolute path of your project.

This may happen if one of your main drives is called "Macintosh HD" or "Macintosh SSD".

You can prevent this by using quotes around the paths.